requirements.txt
bcrypt==3.2.0
blinker==1.4
boto3==1.16.10
boto3-stubs==1.16.12.0
botocore==1.19.10
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
firebase-admin==4.5.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-Script==2.0.6
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.1
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==1.0.2
mysql-connector-python==8.0.21
numpy==1.19.4
Pillow==8.0.1
proto-plus==1.13.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.5.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
rsa==4.7
s3transfer==0.3.3
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
